+++ ding/policy/bc.py
@@ -15,6 +15,8 @@ from ding.utils.data import default_collate, default_decollate
 from ding.rl_utils import q_nstep_td_data, q_nstep_sql_td_error, get_nstep_return_data, get_train_sample
 from ding.worker.collector.interaction_serial_evaluator import InteractionSerialEvaluator
 from ding.utils import POLICY_REGISTRY
+
+
 @POLICY_REGISTRY.register('bc')
 class BehaviourCloningPolicy(Policy):
     config = dict(
@@ -116,6 +118,7 @@ class BehaviourCloningPolicy(Policy):
                 output = to_device(output, 'cpu')
             output = default_decollate(output)
             return {i: d for i, d in zip(data_id, output)}
+
     def _init_collect(self) -> None:
         r"""
         Overview:
